bot/requirements.txt
python-telegram-bot
requests
ujson
urllib3
pymongo
python-dateutil
